Make request_timeout_seconds a whole number of seconds
qdrant-client takes its timeout as an int and rounds a fraction up, so a
fractional value was honored by pymilvus and silently changed for Qdrant.
An int is honored exactly by both, and matches max_retry_interval_seconds
on the embedder and language model configurations.

+++ packages/server/src/memmachine_server/common/vector_store/milvus_vector_store.py
@@ -147,7 +147,7 @@ class MilvusVectorStoreCollection(VectorStoreCollection):
         partition_key: str,
         config: VectorStoreCollectionConfig,
         tracker: OperationTracker,
-        request_timeout_seconds: float,
+        request_timeout_seconds: int,
     ) -> None:
         """Initialize with a Milvus client and collection name."""
         self._client = client
@@ -324,7 +324,7 @@ class MilvusVectorStoreParams(BaseModel):
     Attributes:
         client (MilvusClient): Milvus client instance.
         consistency_level (str): Collection consistency level for newly created collections.
-        request_timeout_seconds (float): Seconds any request to Milvus may take.
+        request_timeout_seconds (int): Seconds any request to Milvus may take.
         metrics_factory (MetricsFactory | None): Metrics factory for collecting usage metrics.
     """
 
@@ -336,7 +336,7 @@ class MilvusVectorStoreParams(BaseModel):
         default="Session",
         description="Milvus consistency level for newly created collections",
     )
-    request_timeout_seconds: float = Field(
+    request_timeout_seconds: int = Field(
         ..., gt=0, description="Seconds any request to Milvus may take"
     )
     metrics_factory: InstanceOf[MetricsFactory] | None = Field(
